bcm-knet: strip vlan tag if received untagged at port
The CPU port will always add a vlan tag, but we can check the header for
the state at reception. If it was received untagged, strip the vlan tag
to forward the packet as received to the kernel.

Different ASICs use different DCBs, which have the information stored at
different places. Add a function checking these, based on the documented
dcb types in OpenBCM/sdk-6.5.24/include/soc/shared/dcbformat/.

For now only support a subset of dcb types which we only care about.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>