bcm-knet: allow marking packets as offloaded
When using KNET interfaces in a bridge, Linux will flood or forward any
packets it sees according to default rules. This can be undesirable when
the packet was already flooded or forwarded by the switch itself, since
this will lead to the packets being duplicated.

To avoid this Linux supports marking packets as forwarding offloaded,
which prevents them from being forwarded in software, so add support for
marking packets as such.

Since this may break use cases where packets are redirected to the
controller, but are expected to be flooded/forwarded in software, make
this an optional flag per interface.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>