bcm-knet: do not mark BPDU packets as offloaded
BPDU packets are trapped to controller and never forwarded, so do not
mark them as offloaded. This allows them to be forwarded/flooded in
software if needed.

Note that this requires configuring the switch ASIC to mark traffic as
BPDU via appropriate L2 cache entries, and there is no inherent
automatic classification.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>