bcm-knet: fix reported tx bytes
Do not include FCS and custom data header in tx byte counts. The custom
header is never transmitted on the wire and consumed by the ASIC, so we
should not count it, and we do not count it for rx bytes.

Likewise, rx and tx bytes should exclude the FCS, so we need to
substract it here as well.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>