KNET: push hardware counters to port netif
Automatically push hw counters to netif interfaces whenever the cached
hw counters get updated.

To accomplish this, add a thread to the SDK that iterates over all
existing netifs and updates the counters if it's a port interface.

To not have to poll, register a extra counters callback that gets called
everytime the SDK counter thread collects counters, and triggers the
knet counter update thread.

Since the software counters are still interesting, delegate them to
IFLA_OFFLOAD_XSTATS_CPU_HIT type offload stats, which seem to be the
most appropriate.

For now only use a subset of counters to keep the size reasonable,
selected by matching the available counters [1] to rtnl_link_stats64
from the kernel [2].

[1] https://github.com/Broadcom-Network-Switching-Software/OpenBCM/blob/master/sdk-6.5.24/include/bcm/stat.h
[2] https://elixir.bootlin.com/linux/v6.6.90/source/include/uapi/linux/if_link.h#L42

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>